moved the reports mock to conftest, fixed some tests
