+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import json
 import os
 from contextlib import contextmanager
-from datetime import date, datetime, timedelta
+from datetime import date, datetime, timedelta, timezone
 from typing import List
 from unittest.mock import Mock
 from uuid import UUID, uuid4
@@ -10,6 +10,7 @@ import pytest
 import requests
 from bs4 import BeautifulSoup
 from flask import Flask, template_rendered, url_for
+from freezegun import freeze_time
 from notifications_python_client.errors import HTTPError
 from notifications_utils.url_safe_token import generate_token
 from pytest_mock import MockerFixture
@@ -4888,3 +4889,48 @@ def create_notifications(
         postage=postage,
         to=to,
     )
+
+
+@freeze_time("2025-01-01 00:01:00.000000")
+def mock_reports_data():
+    return [
+        {
+            "id": "report-1",
+            "status": "ready",
+            "language": "en",
+            "expires_at": (datetime.now(timezone.utc) + timedelta(days=1)).isoformat(),
+            "requested_at": datetime.now(timezone.utc).isoformat(),
+            "requesting_user": {
+                "id": "user-1",
+                "name": "Test User",
+            },
+            "url": "https://example.com/report-1.csv",
+        },
+        {
+            "id": "report-2",
+            "status": "ready",
+            "language": "en",
+            "expires_at": (datetime.now(timezone.utc) - timedelta(days=1)).isoformat(),
+            "requested_at": datetime.now(timezone.utc).isoformat(),
+            "requesting_user": {
+                "id": "user-1",
+                "name": "Test User",
+            },
+        },
+        {
+            "id": "report-3",
+            "status": "generating",
+            "language": "fr",
+            "expires_at": (datetime.now(timezone.utc) + timedelta(days=1)).isoformat(),
+            "requested_at": datetime.now(timezone.utc).isoformat(),
+            "requesting_user": {
+                "id": "user-1",
+                "name": "Test User",
+            },
+        },
+    ]
+
+
+@pytest.fixture(scope="function")
+def mock_get_reports(mocker):
+    return mocker.patch("app.reports_api_client.get_reports_for_service", return_value=mock_reports_data())
